First setup of a node webserver and making request
Making simple request with the http-parser and running through this debugging to see how it worked.

Make request to ‘localhost:1337’ once the app.js is run in terminal to make the “Hello world!” appear in the browser. This is something that came from my webserver and put it in the browser.

This is the begining of real things to come in the course!