Sending HTML data over our webserver to 1337
This code allows us to send the contents of index.html over our web server to the web browser.

This means that the changes we make in our html will alwasy be fresh on the web server.

It also shows the use of templates and how they can be replaced with variables in the Node/javascript side of things.

All run fine.