Express, Nodemon, & NPM Init installed and used
Express, Nodemon, & NPM Init installed and used.

Expres is usd to make the same webserver code as before but a lot easier.

Nodemon should be used from now on to auto reload the Node app.js when changes are made to it’s code.

Init or packages.json in the folder coatains the dependancies for this code. `npm update` to update them.